requirements_utf8.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
cssselect2==0.8.0
dj-database-url==3.1.0
Django==5.2.10
django-recaptcha==4.1.0
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
pillow==12.1.0
pycairo==1.29.0
pycparser==3.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.6.2
python-bidi==0.6.7
python-dotenv==1.2.1
PyJWT==2.8.0
PyYAML==6.0.3
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.11.0
gunicorn==23.0.0
xhtml2pdf==0.2.17
